Extract TargetActorHelper to its own module
